requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-admin-sortable2==1.0.2
django-ckeditor==6.1.0
django-currencies==0.10.1
django-easy-select2==1.5.7
django-import-export==2.6.1
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-mptt==0.13.4
django-rosetta==0.9.7
et-xmlfile==1.1.0
idna==3.3
jsonfield==2.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.3.2
polib==1.1.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
PyYAML==6.0
requests==2.26.0
singlemodeladmin==0.9
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tablib==3.0.0
toml==0.10.2
Unidecode==1.3.2
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
